refactor: isolate HTTP security checks
